add file existence checking

+++ MDMC/exporters/exporter.py
@@ -1,5 +1,5 @@
 """Module for exporter abstract class"""
-
+import os.path
 from abc import ABC, abstractmethod
 
 from MDMC.common.decorators import repr_decorator
@@ -31,8 +31,10 @@ class Exporter(ABC):
         """
         # pylint: disable=consider-using-with
         # as this is an abstracted open method
-
-        self.file = open(self.file_name, 'w', encoding='UTF-8')
+        if os.path.exists(self.file_name):
+            self.file = open(self.file_name, 'w', encoding='UTF-8')
+        else:
+            self.file = open(self.file_name, 'x', encoding='UTF-8')
 
     def __exit__(self, exception_type, exception_value, traceback) -> None:
         """Closes the open file after parsing"""
